including validations: email and required fields

+++ SDK/ProfileSDK/Sources/3InterfaceAdapter/Presenters/SignIn/Impl/SignInPresenterImpl.swift
@@ -18,15 +18,18 @@ public class SignInPresenterImpl: SignInPresenter  {
     private let saveKeyChainEmailUseCase: SaveKeyChainRememberEmailUseCase
     private let getKeyChainEmailUseCase: GetKeyChainRememberEmailUseCase
     private let delKeyChainEmailUseCase: DeleteKeyChainRememberEmailUseCase
+    private let emailValidator: EmailValidations
     
     public init(authUseCase: AuthenticateUseCase,
                 saveKeyChainEmailUseCase: SaveKeyChainRememberEmailUseCase,
                 getKeyChainEmailUseCase: GetKeyChainRememberEmailUseCase,
-                delKeyChainEmailUseCase: DeleteKeyChainRememberEmailUseCase ) {
+                delKeyChainEmailUseCase: DeleteKeyChainRememberEmailUseCase,
+                emailValidator: EmailValidations ) {
         self.authUseCase = authUseCase
         self.saveKeyChainEmailUseCase = saveKeyChainEmailUseCase
         self.getKeyChainEmailUseCase = getKeyChainEmailUseCase
         self.delKeyChainEmailUseCase = delKeyChainEmailUseCase
+        self.emailValidator = emailValidator
     }
 
     public func getEmailKeyChain() -> String? {
@@ -38,17 +41,54 @@ public class SignInPresenterImpl: SignInPresenter  {
         return nil
     }
     
+    private func validations(email: String, password: String) -> String? {
+        
+        if let failMsg = isValidRequiredFields(email: email, password: password) {
+            return failMsg
+        }
+        
+        return isValidEmail(email: email)
+    }
+    
+    private func isValidEmail(email: String) -> String? {
+        if !emailValidator.validate(email: email) {
+            return "\nO email está fora do padrão \n"
+        }
+        return nil
+    }
+    
+    private func isValidRequiredFields(email: String, password: String) -> String? {
+        var fieldFails = [String]()
+        
+        if email.isEmpty {fieldFails.append("E-mail") }
+        
+        if password.isEmpty { fieldFails.append("Senha") }
+        
+        var failMsg: String?
+        if (fieldFails.count > 1) {
+            failMsg = "Os campos: \(fieldFails.description), devem estar preenchidos"
+        }
+        if (fieldFails.count == 1) {
+            failMsg = "O campo: \(fieldFails.description), deve estar preenchido"
+        }
+        
+        return failMsg
+    }
+    
     public func login(email: String, password: String, rememberPassword: Bool = false) {
+        if let msgInvalid = validations(email: email, password: password) {
+            outputDelegate?.error(msgInvalid)
+            return
+        }
+        
         Task {
             do {
                 let userId = try await authUseCase.emailPasswordAuth(email: email, password: password)
                 
                 try delKeyChainEmailUseCase.delete()
-                
                 if rememberPassword {
                     saveRememberEmail(email)
                 }
-                
                 DispatchQueue.main.async { [weak self] in
                     self?.outputDelegate?.success(userId)
                 }
@@ -56,7 +96,6 @@ public class SignInPresenterImpl: SignInPresenter  {
                 DispatchQueue.main.async { [weak self] in
                     self?.outputDelegate?.error("Email ou Senha inválidos")
                 }
-                
             }
         }
